Added a Characteristics object to the JSON definition allowing users to define the behavior of the datacenter itself in a simulation.
Added a function to the HostClass object that generates a CloudSim Host
from the current state of the HostClass object.

Began work on a JsonParser class that reads in a JSON file from the hard
drive and will turn it into its CloudSim counterparts ready to be ran
in a simulation.