Fixed a bug in the DAMParse (formerly known as DependencyFactory) that was omitting the final child of the file. Now works properly.
Created a function that randomly generates nxn dependency matrix and stores
it in a .dam file. Every generated file is guarenteed not to have any self-
dependent cloudlets.

Changed DAMParse (formely known as DependencyFactory) to be an object
instead of a collection of static functions for consistency with DAMSmith
and JSONSmith/JSONParse.