Implemented a Workflow class that keeps track of the Cloudlets for a workflow as well as the weighted dependencies between those Cloudlets (using the WeightedWorkflowMapping class).
Implemented a simple WorkflowDatacenter wich represents a Datacenter capable of accepting
and processing Workflows.

Implemented a simple WorkflowBroker that accepts a Workflow and schedules it to an available
Workflow Datacenter. This class will eventually be used to intellegently match a Workflow
to a Datacenter selected from a federation of datacenters.

Created a SIUTags file that will keep track of all the message Tags that will be passed via
CloudSim's message passing interface.

Currently have the WorkflowBroker sending a Workflow to a WorkflowDatacenter during the
simulation.